--- a/yara/yarascan/yarascan.py
+++ b/yara/yarascan/yarascan.py
@@ -21,11 +21,13 @@ Process a payload using yara
 
 """
 
+import asyncio
+import functools
 import os
 import yara
 
 from pathlib import Path
-from typing import Dict, Generator
+from typing import Dict, List
 from inspect import currentframe, getframeinfo
 
 from stoq.helpers import StoqConfigParser
@@ -65,7 +67,7 @@ class YaraPlugin(WorkerPlugin, DispatcherPlugin):
     async def scan(self, payload: Payload, request: Request) -> WorkerResponse:
         results = {
             'matches': [
-                m for m in self._yara_matches(payload.content, self.worker_rules)
+                m for m in await self._yara_matches(payload.content, self.worker_rules)
             ]
         }
         return WorkerResponse(results=results)
@@ -74,7 +76,7 @@ class YaraPlugin(WorkerPlugin, DispatcherPlugin):
         self, payload: Payload, request: Request
     ) -> DispatcherResponse:
         dr = DispatcherResponse()
-        for match in self._yara_matches(payload.content, self.dispatch_rules):
+        for match in await self._yara_matches(payload.content, self.dispatch_rules):
             if match['meta'].get('save', '').lower().strip() == 'false':
                 payload.results.payload_meta.should_archive = False
             plugin_names = self._extract_plugin_names(match)
@@ -94,16 +96,22 @@ class YaraPlugin(WorkerPlugin, DispatcherPlugin):
         else:
             return yara.compile(filepath=filepath)
 
-    def _yara_matches(self, content: bytes, rules: yara) -> Generator[Dict, None, None]:
-        matches = rules.match(data=content, timeout=self.timeout)
+    async def _yara_matches(self, content: bytes, rules: yara) -> List[Dict]:
+        loop = asyncio.get_event_loop()
+        matches = await loop.run_in_executor(
+            None,
+            functools.partial(rules.match, data=content, timeout=self.timeout)
+        )
+        out_list = []
         for match in matches:
-            yield {
+            out_list.append({
                 'tags': match.tags,
                 'namespace': match.namespace,
                 'rule': match.rule,
                 'meta': match.meta,
                 'strings': match.strings[: self.strings_limit],
-            }
+            })
+        return out_list
 
     def _extract_plugin_names(self, match: dict) -> set:
         plugin_names = set()