--- a/yara/yarascan/yarascan.py
+++ b/yara/yarascan/yarascan.py
@@ -27,7 +27,7 @@ import os
 import yara
 
 from pathlib import Path
-from typing import Dict, List
+from typing import AsyncGenerator, Dict
 from inspect import currentframe, getframeinfo
 
 from stoq.helpers import StoqConfigParser
@@ -67,7 +67,7 @@ class YaraPlugin(WorkerPlugin, DispatcherPlugin):
     async def scan(self, payload: Payload, request: Request) -> WorkerResponse:
         results = {
             'matches': [
-                m for m in await self._yara_matches(payload.content, self.worker_rules)
+                m async for m in self._yara_matches(payload.content, self.worker_rules)
             ]
         }
         return WorkerResponse(results=results)
@@ -76,7 +76,7 @@ class YaraPlugin(WorkerPlugin, DispatcherPlugin):
         self, payload: Payload, request: Request
     ) -> DispatcherResponse:
         dr = DispatcherResponse()
-        for match in await self._yara_matches(payload.content, self.dispatch_rules):
+        async for match in self._yara_matches(payload.content, self.dispatch_rules):
             if match['meta'].get('save', '').lower().strip() == 'false':
                 payload.results.payload_meta.should_archive = False
             plugin_names = self._extract_plugin_names(match)
@@ -96,22 +96,20 @@ class YaraPlugin(WorkerPlugin, DispatcherPlugin):
         else:
             return yara.compile(filepath=filepath)
 
-    async def _yara_matches(self, content: bytes, rules: yara) -> List[Dict]:
+    async def _yara_matches(self, content: bytes, rules: yara) -> AsyncGenerator[Dict, None]:
         loop = asyncio.get_event_loop()
         matches = await loop.run_in_executor(
             None,
             functools.partial(rules.match, data=content, timeout=self.timeout)
         )
-        out_list = []
         for match in matches:
-            out_list.append({
+            yield {
                 'tags': match.tags,
                 'namespace': match.namespace,
                 'rule': match.rule,
                 'meta': match.meta,
                 'strings': match.strings[: self.strings_limit],
-            })
-        return out_list
+            }
 
     def _extract_plugin_names(self, match: dict) -> set:
         plugin_names = set()