The YARA match function is blocking, this pull wraps it in an event loop.

To test it, I used the following to simulate a slow YARA match.
```
    def _blocking_yara(self, rules, content):
        time.sleep(1)
        return rules.match(data=content, timeout=self.timeout)

```

Would it be better prefer to include a decorator in the stoq utils to replicate this behavior in situations where the plugin cannot be converted to support asyncio?  
```
def asyncio_task_executor(func, *, loop=None, executor=None):
    @functools.wraps(func)
    async def wrapper(*args, **kwargs):
        _loop = loop if loop is not None else asyncio.get_event_loop()
        return await _loop.run_in_executor(
            executor=executor,
            func=functools.partial(func, *args, **kwargs)
        )
    return wrapper
```