@malvidin 2020-06-08
Times from running `time stoq run` on 100 small files:
```
# master
real    0m11.193s
user    0m27.326s
sys     0m2.391s

# master with 0.1 second delay:
real    0m40.609s
user    0m22.279s
sys     0m1.989s


# async_yara
real    0m11.295s
user    0m27.463s
sys     0m2.488s

# async_yara with 0.1 second delay:
real    0m12.015s
user    0m26.457s
sys     0m2.412s


```