@mlaferrera 2020-06-18
Again, thanks @malvidin! Your help is greatly appreciated! I'm seeing about a drastic improvement on performance similar to your test cases above.

I've updated it slightly to leverage AsyncGenerators for efficiency. Please take a look and let me know your thoughts. If you are good with it, I'll go ahead and merge.