@malvidin 2020-06-19
I prefer the change to use AsyncGenerators.
 
I am concerned about the consistency of using executors. Currently in this pull request, it uses executor that uses the default thread pool.
The hash plugin, which is CPU bound after the Payload is loaded into memory, does not run in a process pool.
The exif plugin, which is also CPU bound, is executed with a subprocess.

I do not see an option similar/related to provider_consumers that defines how many processes and threads can be assigned, and I have not checked to see if subprocess execution will impact any thread/process configuration.