@mlaferrera 2020-07-08
Perhaps I am misunderstanding something, I don't quite understand your concern about the consistency of using executors. Admittedly, it is probably better practice to not use the default thread pool, so I can go ahead and update that to leverage `concurrent.futures.ThreadPoolExecutor`. Objects aren't modified within the plugin, but rather returned to the core framework to be updated with their response which will maintain consistency of results. Do you mind elaborating on your concern a bit?