@mlaferrera 2020-07-27
Thanks, again @malvidin. I've run a few tests, and it looks like this version is considerably faster. I will be merging it shortly.

Old:
```
________________________________________________________
Executed in    6.01 secs   fish           external 
   usr time    5.62 secs  432.00 micros    5.62 secs 
   sys time    0.39 secs   52.00 micros    0.39 secs 
```

New:
```
________________________________________________________
Executed in    3.47 secs   fish           external 
   usr time    5.70 secs  383.00 micros    5.70 secs 
   sys time    0.42 secs   46.00 micros    0.42 secs 
```